+++ CHANGELOG.md
@@ -15,7 +15,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 # [0.11.2] - 2022-10-05
 ### Fixed
- - "No Codable Characters" filtered out of the Rep Score Progress plots
+ - "No Codeable Characters" filtered out of the Rep Score Progress plots
  - "Logout" button sets up for a successful next login done in the same session
  - Date sort for Rep Score Progress plots and Qualitative Notes
  - Font size for labels in color map plots